data-loaders/requirements.txt
mem0ai>=0.1.0
graphiti-core[google-genai]>=0.28.0
cognee>=0.1.0
anthropic>=0.30.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
